sets remix values regardless of mode
